+++ .swarm-state.md
@@ -186,3 +186,13 @@ With 20+ Knob instances on screen:
 ## Status: complete
 
 **Summary**: The audit found that `MagicKnob.tsx` — despite being orphaned in production — contains severe latent performance issues: per-instance GPU device creation, per-instance shader compilation, unconditional RAF loops with no dirty-checking, and missing GPU teardown. If mounted, even a single instance would waste significant GPU/CPU resources. The production knob (`Knob.tsx`) is actually well-behaved: it uses compositor-only `transform` animations, is memoized, and does not force layout reflow. The primary DOM-side risk is minor listener churn during drag interactions. `useStableKnobConfig` is dead code for the main Knob.tsx consumers but correctly serves `HardwareModule`. A minimal `engineTelemetry.recordLatency('knob-raf', ...)` probe has been added to `MagicKnob.tsx` so that if it is ever mounted, the Engine HUD (`Ctrl+Shift+E`) will surface per-frame render timing.
+
+
+---
+
+[step 3 — 2026-05-11T20:53:44+02:00]
+Status: complete
+Files modified: src/components/KnobGPUContext.ts (new), src/components/MagicKnob.tsx (refactored)
+What was done: Created KnobGPUContext.ts singleton that lazily initializes one shared GPUAdapter, GPUDevice, and GPURenderPipeline, maintains a Set of registered knob entries each with its own canvas context, uniformBuffer, and bindGroup, and drives all knobs from a single requestAnimationFrame loop. Refactored MagicKnob.tsx to remove all per-instance WebGPU initialization (requestAdapter, requestDevice, pipeline creation, RAF loop) and replaced it with a minimal useEffect calling registerKnob(canvas, () => (valueRef.current - min) / (max - min)). All interaction logic, canvas markup, bezel image, label, and aria attributes are preserved unchanged. The WGSL shader was relocated verbatim into KnobGPUContext.ts.
+Next step: none
+Blockers: none
